changed main page to search. fixed bug with registration.
